Limpiando codigo de UI y User y eliminando newPassword
