+++ index.html
@@ -11,38 +11,41 @@
 </head>
 <body>
     <!-- Navbar principal -->
-    <nav class="navbar navbar-expand-lg navbar-light bg-light style-navbar">
-        <a class="navbar-brand" href="./">
-            <img src="https://i.imgur.com/zg0WCUM.png" class="d-inline-block align-top" alt=":(" loading="lazy">            
-        </a>
-        <div class="collapse navbar-collapse style-navbar-container" id="navbarNav">
-            <ul class="navbar-nav">
-              <li class="nav-item active style-links-navbar">
-                <a class="nav-link-left" href="#">
-                    <i class="fas fa-home"></i>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light style-navbar d-flex justify-content-between">
+        <!-- <a class="navbar-brand" href="./">
+            <img src="https://i.imgur.com/zg0WCUM.png" class="d-inline-block align-top mr-5" alt=":(" loading="lazy">            
+        </a> -->
+        <div>
+            <ul class="list-header">
+              <li class="nav-item active style-links-navbar d-flex align-items-center mr-4">
+                <a class="nav-link-left ml-2 text-decoration-none position-relative d-flex" href="#">
+                    <h6 class="">Home</h6>
+                    <i class="fas fa-home position-absolute absolute-icon"></i>
                 </a>
               </li>
-              <li class="nav-item style-links-navbar">
-                <a id="peers" class="nav-link-left" href="#">
-                    <i class="fas fa-user-friends"></i>
+              <li class="nav-item style-links-navbar d-flex align-items-center mr-4 ml-4">
+                <a id="peers" class="nav-link-left position-relative text-decoration-none d-flex" href="#">
+                    <h6 class="">Advisors</h6>
+                    <i class="fas fa-user-friends position-absolute absolute-icon more-right"></i>
                 </a>
               </li>
-              <li class="nav-item style-links-navbar">
-                <a class="nav-link-left" href="#">
-                    <i class="fas fa-star"></i>
+              <li class="nav-item style-links-navbar d-flex align-items-center ml-4">
+                <a class="nav-link-left position-relative text-decoration-none d-flex" href="#">
+                    <h6 class="">Skills</h6>
+                    <i class="fas fa-star position-absolute absolute-icon icon-star"></i>
                 </a>
               </li>
             </ul>
         </div>
         <div>
-            <ul class="navbar-nav">
+            <ul class="list-header">
                 <li class="nav-item active style-links-navbar">
-                    <a id="note" class="nav-link" href="#">
+                    <a id="note" class="mr-3" href="#">
                         <i class="fas fa-bell"></i>
                     </a>
                 </li>
                 <li class="nav-item active style-links-navbar">
-                    <a id="button-profile" type="button" class="nav-link" href="#">
+                    <a id="button-profile" type="button" class="" href="#">
                         <i class="fas fa-user-circle"></i>
                     </a>
                 </li>
