+++ views/login.html
@@ -7,20 +7,20 @@
     <link rel="stylesheet" href="../css/style.css">
     <title>Inicio de sesion - Matter-App</title>
 </head>
-<body class="text-center">
+<body class="text-center img-url">
     <div class="container">
         <form action="/" class="form-login" id="login">
             <div class="row justify-content-center">
-                <div class="col-md-12 col-lg-12 col-xl-12 col-xs-12">
-                    <img class="mb-5 mt-5" src="" alt="Logo">
-                    <h4 class="pb-3">Matter-App - Inicio de sesion</h4>
+                <div class="col-md-12 col-lg-12 col-xl-12 col-xs-12 bg-white rounded-10px mt-5 width-30 p-4">
+                    <!-- <img class="mb-5 mt-5" src="" alt="Logo"> -->
+                    <h4 class="pb-3 mt-3">Log In</h4>
                     <div class="form-group">
-                      <label for="email">Ingresa tu correo</label>
-                      <input type="email" class="form-control email-input" name="email" id="email" placeholder="@tucorreo.com" required>
-                      <label for="email">Ingresa tu contraseña</label>
+                      <label for="email">Enter Your Email</label>
+                      <input type="email" class="form-control email-input" name="email" id="email" placeholder="@youremail.com" required>
+                      <label for="email">Enter Your Password</label>
                       <input type="password" class="form-control password-input" name="password" id="password" placeholder="**************" required>
-                      <button type="submit" class="btn btn-primary mt-3 btn-login">Inicio de sesion</button>
-                      <a type="button" class="btn btn-secondary mt-3" href="./register.html">¿Necesitas registrarte?</a>
+                      <button type="submit" class="btn btn-primary mt-3 btn-login">Log In</button>
+                      <a type="button" class="btn btn-secondary mt-3" href="./register.html">Sign In</a>
                     </div>
                 </div>
             </div>
