+++ views/register.html
@@ -10,17 +10,18 @@
 <body id="body-register">
     <div class="container-form-register">
         <form id="register">
+            <h2 class="mb-4">Sign Up</h2>
             <div class="form-group">
-                <label for="name">Name</label>
-                <input type="text" class="form-control" id="name">
+                <!-- <label for="name">Name</label> -->
+                <input type="text" class="form-control" id="name" placeholder="Enter Name">
             </div>
             <div class="form-group">
-            <label for="email">Email</label>
-            <input type="text" class="form-control" id="email">
+            <!-- <label for="email">Email</label> -->
+            <input type="email" class="form-control" id="email" placeholder="Enter Email">
             </div>
             <div class="form-group">
-            <label for="password">Password</label>
-            <input type="password" class="form-control" id="password">
+            <!-- <label for="password">Password</label> -->
+            <input type="password" class="form-control" id="password" placeholder="Enter Password">
             </div>
             <button type="submit" class="btn btn-primary">Register</button>
         </form>
